# src/swaggerDef.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "E-commerce API",
    "description": "API para un e-commerce deportivo con funcionalidades de autenticación, productos, compras, y más.",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/productos": {
      "get": {
        "summary": "Obtener todos los productos",
        "tags": [
          "Productos"
        ],
        "responses": {
          "200": {
            "description": "Lista de productos obtenida exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Producto"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error al obtener productos"
          }
        }
      }
    },
...
    "/productos/update/all-imagenes": {
      "put": {
        "summary": "Actualizar imágenes de múltiples productos",
        "description": "Actualiza la imagen de los productos en la base de datos a partir de un listado de productos con sus IDs e imágenes.",
        "operationId": "updateALLIMG",
        "tags": [
          "Productos"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "description": "El ID del producto a actualizar.",
                      "example": 52
                    },
                    "imagen_url": {
                      "type": "string",
                      "description": "La nueva URL de la imagen del producto.",
                      "example": "https://i.ibb.co/NyYnpBz/remera-adizero-running.jpg"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Las imágenes de los productos se han actualizado correctamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Imagenes actualizadas correctamente"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error en la actualización de uno o más productos.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "No se pudo actualizar el producto con ID: 52"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor al intentar actualizar los productos.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Error al actualizar productos"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/categorias": {
      "get": {
        "summary": "Obtener todas las categorías",
        "operationId": "getCategorias",
        "tags": [
          "Categorías"
        ],
        "responses": {
          "200": {
            "description": "Lista de categorías obtenida correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Categoria"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
    "/compra": {
      "post": {
        "summary": "Crear una compra",
        "operationId": "crearCompra",
        "tags": [
          "Compras"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Compra"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Compra registrada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "compraId": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Faltan datos obligatorios"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
...
    "/tipos-usuarios": {
      "get": {
        "summary": "Obtener todos los tipos de usuarios",
        "operationId": "getTiposUsuarios",
        "tags": [
          "Tipos de Usuario"
        ],
        "responses": {
          "200": {
            "description": "Tipos de usuarios obtenidos correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TipoUsuario"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Producto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID del producto"
          },
          "nombre": {
            "type": "string",
            "description": "Nombre del producto"
          },
          "descripcion": {
            "type": "string",
            "description": "Descripción del producto"
          },
          "precio": {
            "type": "number",
            "format": "float",
            "description": "Precio del producto"
          },
          "idCategoria": {
            "type": "integer",
            "description": "ID de la categoría del producto"
          }
        },
        "required": [
          "id",
          "nombre",
          "precio"
        ]
      },
      "Categoria": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID de la categoría"
          },
          "nombre": {
            "type": "string",
            "description": "Nombre de la categoría"
          }
        },
        "required": [
          "id",
          "nombre"
        ]
      },
      "Compra": {
        "type": "object",
        "properties": {
          "id_usuario": {
            "type": "integer",
            "description": "ID del usuario que realiza la compra"
          },
          "productos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductoCompra"
            },
            "description": "Lista de productos comprados"
          },
          "monto_total": {
            "type": "number",
            "format": "float",
            "description": "Monto total de la compra"
          }
        },
        "required": [
          "id_usuario",
          "productos",
          "monto_total"
        ]
      },
      "ProductoCompra": {
        "type": "object",
        "properties": {
          "id_producto": {
            "type": "integer",
            "description": "ID del producto"
          },
          "cantidad": {
            "type": "integer",
            "description": "Cantidad del producto comprado"
          }
        },
        "required": [
          "id_producto",
          "cantidad"
        ]
      },
...
      "TipoUsuario": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID del tipo de usuario"
          },
          "nombre": {
            "type": "string",
            "description": "Nombre del tipo de usuario"
          }
        },
        "required": [
          "id",
          "nombre"
        ]
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
}